[Add] 휴식 기능 추가, 아이템 판매 기능 추가

+++ Player.cs
@@ -7,7 +7,7 @@
         public int AttackPower { get; set; }
         public int Defense { get; set; }
         public int Hp { get; set; }
-        public int Gold { get; set; }
+        public double Gold { get; set; }
 
         public Player()
         {
